+++ server/apps/memory-app/src/main/java/org/apache/james/MemoryJamesConfiguration.java
@@ -27,7 +27,7 @@ import org.apache.commons.configuration2.ex.ConfigurationException;
 import org.apache.james.data.UsersRepositoryModuleChooser;
 import org.apache.james.filesystem.api.FileSystem;
 import org.apache.james.filesystem.api.JamesDirectoriesProvider;
-import org.apache.james.jmap.draft.JMAPModule;
+import org.apache.james.jmap.JMAPModule;
 import org.apache.james.server.core.JamesServerResourceLoader;
 import org.apache.james.server.core.MissingArgumentException;
 import org.apache.james.server.core.configuration.Configuration;
