+++ server/container/guice/protocols/jmap/src/main/java/org/apache/james/jmap/MessageIdProbe.java
@@ -17,7 +17,7 @@
  * under the License.                                           *
  ****************************************************************/
 
-package org.apache.james.jmap.draft;
+package org.apache.james.jmap;
 
 import java.util.List;
