+++ server/protocols/jmap-draft-integration-testing/jmap-draft-integration-testing-common/src/test/java/org/apache/james/jmap/draft/methods/integration/SetMessagesMethodTest.java
@@ -85,8 +85,8 @@ import org.apache.james.jmap.AccessToken;
 import org.apache.james.jmap.HttpJmapAuthentication;
 import org.apache.james.jmap.JmapCommonRequests;
 import org.apache.james.jmap.MessageAppender;
-import org.apache.james.jmap.draft.JmapGuiceProbe;
-import org.apache.james.jmap.draft.MessageIdProbe;
+import org.apache.james.jmap.JmapGuiceProbe;
+import org.apache.james.jmap.MessageIdProbe;
 import org.apache.james.junit.categories.BasicFeature;
 import org.apache.james.mailbox.DefaultMailboxes;
 import org.apache.james.mailbox.FlagsBuilder;
