+++ server/protocols/jmap-draft-integration-testing/jmap-draft-integration-testing-common/src/test/java/org/apache/james/jmap/draft/methods/integration/SetVacationResponseTest.java
@@ -39,7 +39,7 @@ import org.apache.james.jmap.AccessToken;
 import org.apache.james.jmap.api.model.AccountId;
 import org.apache.james.vacation.api.Vacation;
 import org.apache.james.vacation.api.VacationPatch;
-import org.apache.james.jmap.draft.JmapGuiceProbe;
+import org.apache.james.jmap.JmapGuiceProbe;
 import org.apache.james.junit.categories.BasicFeature;
 import org.apache.james.util.ValuePatch;
 import org.apache.james.utils.DataProbeImpl;
