+++ server/protocols/webadmin-integration-test/memory-webadmin-integration-test/src/test/java/org/apache/james/webadmin/integration/memory/MemoryUserDeletionIntegrationTest.java
@@ -51,7 +51,7 @@ import org.apache.james.jmap.api.model.PushSubscription;
 import org.apache.james.jmap.api.model.PushSubscriptionCreationRequest;
 import org.apache.james.jmap.api.model.PushSubscriptionServerURL;
 import org.apache.james.jmap.api.pushsubscription.PushSubscriptionRepository;
-import org.apache.james.jmap.draft.JmapGuiceProbe;
+import org.apache.james.jmap.JmapGuiceProbe;
 import org.apache.james.mailbox.model.MailboxACL;
 import org.apache.james.mailbox.model.MailboxPath;
 import org.apache.james.modules.ACLProbeImpl;
@@ -75,9 +75,7 @@ import com.google.inject.multibindings.Multibinder;
 import io.restassured.specification.RequestSpecification;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
-import scala.None;
 import scala.Option;
-import scala.collection.immutable.Seq;
 
 class MemoryUserDeletionIntegrationTest {
     public static class MemoryUserDeletionIntegrationTestProbe implements GuiceProbe {
